+++ .env
@@ -1,4 +1,4 @@
-PYTHONPATH=/app/safe_push_service
+PYTHONPATH=/app/safe_notification_service
 DJANGO_SETTINGS_MODULE=config.settings.local
 C_FORCE_ROOT=true
 DEBUG=False
