+++ docker/web/celery/worker/run.sh
@@ -2,4 +2,4 @@
 
 set -euo pipefail
 
-celery -A safe_push_service.taskapp worker -l INFO
+celery -A safe_notification_service.taskapp worker -l INFO
