+++ safe_notification_service/safe/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class SafeConfig(AppConfig):
-    name = 'safe_push_service.safe'
+    name = 'safe_notification_service.safe'
