+++ safe_notification_service/safe/tasks.py
@@ -2,10 +2,10 @@ from celery import app
 from celery.utils.log import get_task_logger
 from django.conf import settings
 
-from safe_push_service.firebase.client import FirebaseClient, MockedClient
+from safe_notification_service.firebase.client import FirebaseClient, MockedClient
 
 logger = get_task_logger(__name__)
-oid = 'SAFE_PUSH_SERVICE'
+oid = 'SAFE_NOTIFICATION_SERVICE'
 
 
 firebase_client = None
