+++ safe_notification_service/safe/tests/test_serializers.py
@@ -5,8 +5,8 @@ from django.utils import timezone
 from faker import Faker
 from rest_framework.exceptions import ValidationError
 
-from safe_push_service.ether.signing import EthereumSignedMessage
-from safe_push_service.ether.tests.factories import get_eth_address_with_key
+from safe_notification_service.ether.signing import EthereumSignedMessage
+from safe_notification_service.ether.tests.factories import get_eth_address_with_key
 
 from ..serializers import (AuthSerializer, GoogleInAppPurchaseSerializer,
                            NotificationSerializer, PairingDeletionSerializer,
