+++ safe_notification_service/safe/tests/test_validators.py
@@ -1,7 +1,7 @@
 from django.core.exceptions import ValidationError
 from django.test import TestCase
 
-from safe_push_service.ether.tests.factories import get_eth_address_with_key
+from safe_notification_service.ether.tests.factories import get_eth_address_with_key
 
 from ..validators import validate_checksumed_address
