+++ safe_notification_service/utils/singleton.py
